Polish/fix comments and simplify state
We do not need to keep the next line in the _head_line variable, the
entire state is already stored in the _char_iter iterator.

In addition, it is simpler to initialize _last_processed_state to None
rather than to some special value, and just initialize it the first time
we load a line.