Introduce dedicated EOF token; clarify syntax
The explicit EOF token at the end makes parsing a bit simpler. Errors
triggered by the end of input now also point to the column _after_ the
last character, which is a bit clearer.

This commit also clarifies the syntax. The top level element is now
called "Program" and it has zero or one expressions.