Structured diagnostics representation throughout
Prior to this commit, the diagnostics information was passed as a string
through most of the stack, inside exception messages. This commit
changes this, the diagnostics information is now kept in structured
form, as a Diagnostics object, in the exceptions, and only the
user-facing code (see e.g. main.py) is responsible to convert it to a
string.

This currently has the drawback that only a single position is
supported; I had to remove the nice message for missing closing
parenthesis that points to the opening parenthesis. I have to think
about how to better model errors, maybe I need specialized exception
classes.

Another issue is that the diagnostics don't point to the right token in
some cases, see FIXME.